requirements.txt
flake8==4.0.1
mccabe==0.6.1
numpy==1.22.3
pycodestyle==2.8.0
pyflakes==2.4.0
PyOpenGL==3.1.6
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.10.1
winshell==0.6
